requirements.txt
tensorflow
mysql-connector-python==8.0.28
cryptography==37.0.2